example/debugserver: refactor to support command-line model selection and custom address
- Added command-line flags for specifying the model and listen address in main.go.
- Updated README.md with new usage instructions for starting the server with different models and ports.
- Introduced tools.go for calculator and time functionalities, enhancing the server's capabilities.
